I dont know what is in here

+++ Assets/Scripts/Resources.cs
@@ -7,7 +7,7 @@ public class Resources
     public int Wood = 1;
     public int Steel = 1;
     public int Gold = 1;
-    public int Happiness = 100;
+    public float Happiness = 100f;
     public int Population = 5;
     public int PopulationAtWork = 2;
     public int RiverSize = 3;
